requirements.txt
attrs==22.1.0
colorama==0.4.5
iniconfig==1.1.1
numpy==1.23.4
packaging==21.3
play-tictactoe==0.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
